+++ xpeapp_ios/XpeApp/XpeApp/ContentView.swift
@@ -8,6 +8,7 @@
 import SwiftUI
 
 struct ContentView: View {
+    @State var qvstCampaigns: [QvstCampaign] = []
     
     var body: some View {
         TabView {
@@ -23,6 +24,10 @@ struct ContentView: View {
                 .tabItem {
                     Label("Contacts", systemImage: "person.crop.circle.fill")
                 }
+            QvstCampaignsPageView(campaigns: qvstCampaigns)
+                .tabItem {
+                    Label("QVST Temp", systemImage: "questionmark")
+                }
             //Note(Loucas): This is a convenience menu for development purposes
             #if DEBUG && true
             DebugPageView()
@@ -30,6 +35,21 @@ struct ContentView: View {
                     Label("Debug", systemImage: "wrench.fill")
                 }
             #endif
+        }
+        .onAppear {
+            if qvstCampaigns.isEmpty {
+                updateCampaigns()
+            }
+        }
+    }
+    
+    private func updateCampaigns() {
+        Task {
+            guard let cs = await QvstCampaign.fetchActive() else { return }
+            // Note: We have to do UI updates on the main thread to prevent crashes
+            DispatchQueue.main.async {
+                qvstCampaigns = cs
+            }
         }
     }
 }